For protein comparison, a special case in maizecyc is now handled in which a large number of maizecyc proteins did not include the gene model name as one of their synonymes.  This case covers large groups of proteins which reused a nondescript and generic name in the CommonName slot.  We addressed this issue by associated the gene model name from the protein's gene with the protein where possible.
For both protein and gene comparison, identified a filter which users
may wish to dynamically alter, such that Go term association and/or
reaction association can be choosen as the criteria of a sufficiently
annotated gene/protein.  This is not yet implemented, but is hard-coded
to the most likely scenario for each.