A temporary fix as a way to output a list of pathway - gene membership.
For each pathway, list each gene that is associated with a reaction in
that pathway.  Do this for CornCyc and MaizeCyc. Only output pathways
which are common to both CornCyc and MaizeCyc and have at least one gene
associated with that pathway in both databases. We are interested in
showing when CornCyc and MaizeCyc disagree on pathway - gene membership.
(tested on v17.5)